navigation bar added, reroute / to /games
