+++ netlify.toml
@@ -1,3 +1,9 @@
 [build]
 command = "npm run build"
 publish = "build"
+
+[[redirects]]
+from = "https://twoshotslater.com/"
+to = "https://twoshotslater.com/games"
+status = 200
+force = true
