+++ src/routes/about/+page.js
@@ -1,0 +1,1 @@
+
