+++ src/routes/contact/+page.js
@@ -1,0 +1,1 @@
+
