+++ src/routes/games/+page.js
@@ -1,0 +1,1 @@
+
