+++ static/global.css
@@ -15,7 +15,6 @@ p {
 }
 
 a {
-  font-family: "Poppins", "Arial";
   font-size:x-large;
   color: #F1342E;
 }
